+++ BE/health-service/src/main/java/com/s206/health/exercise/repository/ExerciseCategoryRepository.java
@@ -3,5 +3,5 @@ package com.s206.health.exercise.repository;
 import com.s206.health.exercise.entity.ExerciseCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExerciseCategoryRepository extends JpaRepository<ExerciseCategory, Long> {
+public interface ExerciseCategoryRepository extends JpaRepository<ExerciseCategory, Integer> {
 }
